Made buttons to switch between tables functional
