--- petclinic-frontend/src/features/visits/Sidebar.css
@@ -1,108 +1,0 @@
-
-@import './VisitsColors.css';
-
-
-#sidebar{
-    height:100vh;
-    width: 250px;
-    padding: 5px 1em;
-    padding-left: 0;
-    background-color: var(--secondary-clr);
-    box-shadow: inset 0 0 4px 0 rgba(0,0,0,0.25);
-    position: sticky;
-    top: 0;
-    align-self: start;
-    transition: 300ms ease-in-out;
-    overflow: hidden;
-    text-wrap: nowrap;
-    
-}
-
-#sidebar.close{
-    padding: 5px;
-}
-
-ul{
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-ul > li:first-child{
-    padding-left: 20px;
-    margin: 8px;
-}
-
-
-
-#sidebar li > a {
-    flex-grow: 100;
-    text-decoration: none;
-    color: var(--text-clr);
-    display: flex;
-    align-items: center;
-    border-top-right-radius: 15px;
-    border-bottom-right-radius: 15px;
-    padding: 2px;
-    padding-left: 20px;
-}
-
-#sidebar ul> li> a:hover{
-    background-color: var(--hover-clr);
-    cursor: pointer;
-}
-
-#sidebar .active, #sidebar .active:hover{
-    background-color: var(--active-clr);
-    font-weight: bold;
-}
-
-#sidebar .emergency{
-    color: var(--emergency-clr);
-}
-
-#sidebar .activeemergency, #sidebar .activeemergency:hover{
-    color: var(--base-clr);
-    background-color: var(--emergency-clr);
-    font-weight: bold;
-}
-
-
-#sidebar h2{
-    font-weight: bold;
-    margin-bottom: 0px;
-    display: flex;
-    justify-content: space-between;
-    flex-grow: 100;
-}
-
-#sidebar li {
-    display: flex;
-}
-#sidebar button{
-    display: flex;
-    gap: 10px;
-    align-items: center;
-    text-align: left;
-    flex-grow: 100;
-    margin-left: 20px;
-    padding: 8px;
-    padding-left: 16px;
-    font-size: 16px;
-}
-
-#sidebar button svg{
-    width:28px;
-    height:28px;
-    fill:#ffffff;
-}
-/* 
-.unread-icon{
-    width:20px;
-    height:20px;
-    
-    fill:none;
-    
-    stroke:#000000;
-    
-} */
