+++ petclinic-frontend/src/features/visits/VisitListTable.css
@@ -1,5 +1,18 @@
 
-@import './VisitsColors.css';
+:root{
+    --base-clr: #FFFFFF;
+    --secondary-clr: #F5F5FF;
+    --hover-clr: #E1E5EF;
+    --active-clr: #d4d4e7;
+    --dark-clr: #212529;
+    --emergency-clr: #FF3C56;
+    --accent-clr: #5e63ff;
+}
+
+.page-container{
+  display:grid;
+  grid-template-columns: auto 1fr;
+}
 
 .visit-actions {
   display: flex;
@@ -126,5 +139,116 @@ h2 {
   to {
     opacity: 1;
     transform: scale(1);
-  } */
-/* }    */
+  }
+}
+
+
+main{
+  padding: 20px;
+}
+
+
+#sidebar{
+    height:100vh;
+    width: 250px;
+    padding: 5px 1em;
+    padding-left: 0;
+    background-color: var(--secondary-clr);
+    box-shadow: inset 0 0 4px 0 rgba(0,0,0,0.25);
+    position: sticky;
+    top: 0;
+    align-self: start;
+    transition: 300ms ease-in-out;
+    overflow: hidden;
+    text-wrap: nowrap;
+    
+}
+
+#sidebar.close{
+    padding: 5px;
+}
+
+ul{
+    list-style: none;
+    padding: 0;
+    margin: 0;
+}
+
+ul > li:first-child{
+    padding-left: 20px;
+    margin: 8px;
+}
+
+/* Sidebar */
+
+#sidebar li > a {
+    flex-grow: 100;
+    text-decoration: none;
+    color: var(--text-clr);
+    display: flex;
+    align-items: center;
+    border-top-right-radius: 15px;
+    border-bottom-right-radius: 15px;
+    padding: 2px;
+    padding-left: 20px;
+}
+
+#sidebar ul> li> a:hover{
+    background-color: var(--hover-clr);
+    cursor: pointer;
+}
+
+#sidebar .active, #sidebar .active:hover{
+    background-color: var(--active-clr);
+    font-weight: bold;
+}
+
+#sidebar .emergency{
+    color: var(--emergency-clr);
+}
+
+#sidebar .activeemergency, #sidebar .activeemergency:hover{
+    color: var(--base-clr);
+    background-color: var(--emergency-clr);
+    font-weight: bold;
+}
+
+
+#sidebar h2{
+    font-weight: bold;
+    margin-bottom: 0px;
+    display: flex;
+    justify-content: space-between;
+    flex-grow: 100;
+}
+
+#sidebar li {
+    display: flex;
+}
+#sidebar button{
+    display: flex;
+    gap: 10px;
+    align-items: center;
+    text-align: left;
+    flex-grow: 100;
+    margin-left: 20px;
+    padding: 8px;
+    padding-left: 16px;
+    font-size: 16px;
+}
+
+#sidebar button svg{
+    width:28px;
+    height:28px;
+    fill:#ffffff;
+}
+/* 
+.unread-icon{
+    width:20px;
+    height:20px;
+    
+    fill:none;
+    
+    stroke:#000000;
+    
+} */
