--- petclinic-frontend/src/features/visits/VisitsColors.css
@@ -1,9 +1,0 @@
-:root{
-    --base-clr: #FFFFFF;
-    --secondary-clr: #F5F5FF;
-    --hover-clr: #E1E5EF;
-    --active-clr: #d4d4e7;
-    --dark-clr: #212529;
-    --emergency-clr: #FF3C56;
-    --accent-clr: #5e63ff;
-}
