+++ petclinic-frontend/src/pages/Visit/Visit.tsx
@@ -1,5 +1,4 @@
 import VisitListTable from '@/features/visits/VisitListTable';
-import SideBar from '@/features/visits/Sidebar';
 import { NavBar } from '@/layouts/AppNavBar';
 import './Visits.css';
 
@@ -8,8 +7,7 @@ export default function Visits(): JSX.Element {
     <>
       <NavBar />
 
-      <div className="page-container">
-        <SideBar />
+      <div>
         <VisitListTable />
       </div>
     </>
