+++ petclinic-frontend/src/pages/Visit/Visits.css
@@ -1,5 +1,2 @@
 
-.page-container{
-  display:grid;
-  grid-template-columns: auto 1fr;
-}
+
